+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.23.0 - 2023-12-24
+- Find symbol usages from symbol
+- [Find symbol usages from Var](https://github.com/pedrorgirardi/Pep/issues/68)
+- [Symbol highlight should also highlight Vars](https://github.com/pedrorgirardi/Pep/issues/71)
+
 ## 0.22.0 - 2023-12-23
 - `PgPepInspect` now shows data in a popup
 - New Command `PgPepGotoUsageInViewCommand`